% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{buildEloDiffsMatrix}
\alias{buildEloDiffsMatrix}
\title{Build a matrix of elo differences between two teams}
\usage{
buildEloDiffsMatrix(teamElos)
}
\arguments{
\item{teamElos}{A dataframe in the same format as a result from
\code{\link{extractTeamElos}}: team names and elo ratings.}
}
\value{
A dataframe where each row and each column correspond to a team,
and individual cells correspond to the Elo rating of the row-team minus
the Elo rating of the column-team.
}
\description{
Build a matrix of elo differences between two teams
}
\author{
astroud
}
\keyword{internal}
