% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{buildModelsElo}
\alias{buildModelsElo}
\title{Obtain Point Differential and Win Probability Matrices using an Elo model}
\usage{
buildModelsElo(eloLogs, season)
}
\arguments{
\item{eloLogs}{A data frame of game logs including the columns \code{season},
\code{team1}, \code{team2}, \code{elo1_pre}, \code{elo2_pre}, \code{score1}, and
\code{score2}. The output of a call to \code{\link{getEloGameLogs}} will work here.}

\item{season}{An integer indicating the NBA season for which outputs are desired.}
}
\value{
A list of two matrices, where the first matrix is the output of
\code{\link{buildWinProbMatrixElo}} and the second matrix is the output of
\code{\link{buildPtDiffMatrixElo}}.
}
\description{
Obtain Point Differential and Win Probability Matrices using an Elo model
}
\examples{
\dontrun{
data(eloLogs2017)
nbaMatrices = buildModelsElo(eloLogs2017)
}

}
\seealso{
\code{\link{buildModelsBT}} to obtain the same list of matrices using a
Bradley-Terry model instead of Elo ratings.
}
\author{
astroud
}
