% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPtDiffMatrixBT.R
\name{buildPtDiffMatrixBT}
\alias{buildPtDiffMatrixBT}
\title{Predict point differential in a matchup between two teams by
fitting Bradley-Terry model}
\usage{
buildPtDiffMatrixBT(GameLog, regularize = TRUE)
}
\arguments{
\item{GameLog}{A data frame containing individual game records
with the following columns: \code{H.TEAM}, \code{A.TEAM},
\code{H.PTS}, and \code{A.PTS}.}

\item{regularize}{logical. If \code{TRUE} regularization term
will be added to predict team strength. It is recommended but
can be omitted.}
}
\value{
A matrix with each entry representing the point differential in a
matchup of the team in that row at home against the team in that column on the
road. Note that the diagonal terms are estimates of the home court advantage
coefficient.
}
\description{
Predict point differential in a matchup between two teams by
fitting Bradley-Terry model
}
\details{
Note that for the Bradley-Terry models here we based much of our code
off of \url{http://www.stanford.edu/class/stats50/files/5-bradley-terry-complete.r}
and \url{http://www.stanford.edu/class/stats50/files/5-bradley-terry.pdf}.
}
\examples{
buildPtDiffMatrixBT(gamelogs2017_full)
buildPtDiffMatrixBT(gamelogs2017_simple, regularize = FALSE)

}
\seealso{
\code{\link{buildWinProbMatrixBT}} for obtaining a similar matrix with predicted
win probabilities, and \code{\link{buildModelsBT}} to obtain both types of matrices at once.
}
\author{
kitliu5 astroud
}
