% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{buildPtDiffMatrixElo}
\alias{buildPtDiffMatrixElo}
\title{Predict Point Differentials between every pair of NBA teams with an Elo model}
\usage{
buildPtDiffMatrixElo(eloLogs, season)
}
\arguments{
\item{eloLogs}{A dataframe of Elo gamelogs, as from the output of
\code{\link{getEloGameLogs}}.}

\item{season}{An integer corresponding to the NBA season for which outputs are desired.}
}
\value{
A matrix with each entry representing the point differential in a
matchup of the team in that row at home against the team in that column on
the road. Note that the diagonal terms are estimates of the home court advantage
coefficient.
}
\description{
Predict Point Differentials between every pair of NBA teams with an Elo model
}
\examples{
\dontrun{
data(eloLogs2017)
buildPtDiffMatrixElo(eloLogs2017, 2017)
}

}
\seealso{
\code{\link{buildWinProbMatrixElo}} for obtaining a similar matrix
with predicted win probabilities, and \code{\link{buildModelsElo}} to obtain
both types of matrices at once.
}
\author{
astroud
}
