% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWinProbMatrixBT.R
\name{buildWinProbMatrixBT}
\alias{buildWinProbMatrixBT}
\title{Predict wining probability of each team in a matchup by
fitting Bradley-Terry model}
\usage{
buildWinProbMatrixBT(GameLog, regularize = T)
}
\arguments{
\item{GameLog}{A data frame containing individual game records
with the following columns: \code{H.TEAM}, \code{A.TEAM},
\code{H.PTS}, and \code{A.PTS}.}

\item{regularize}{logical. If \code{TRUE} regularization term
will be added to predict team strength. It is recommended but
can be set to \code{FALSE}.}
}
\value{
A probability matrix with each entry representing the
probability of the team in that row defeating the team in that
column in a game where the team in the row is the home team, and
the team in the column is the away team.
}
\description{
Predict wining probability of each team in a matchup by
fitting Bradley-Terry model
}
\details{
Note that diagonal entries of the return matrix are
not exactly 50% because home-court advantage is taken into account.
Also, for the Bradley-Terry models here we based much of our code
off of
\url{http://www.stanford.edu/class/stats50/files/5-bradley-terry-complete.r}
and \url{http://www.stanford.edu/class/stats50/files/5-bradley-terry.pdf}.
}
\examples{
buildWinProbMatrixBT(gamelogs2017_full)
buildWinProbMatrixBT(gamelogs2017_simple)

}
\seealso{
\code{\link{buildPtDiffMatrixBT}} for obtaining a similar matrix with predicted
point differentials, and \code{\link{buildModelsBT}} to obtain both types of matrices at once.
}
\author{
kitliu5 and astroud
}
