% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{buildWinProbMatrixElo}
\alias{buildWinProbMatrixElo}
\title{Predict Win Probabilities between every pair of NBA teams with an Elo model}
\usage{
buildWinProbMatrixElo(eloLogs, season)
}
\arguments{
\item{eloLogs}{A dataframe of Elo gamelogs, as from the output of
\code{\link{getEloGameLogs}}.}

\item{season}{An integer corresponding to the NBA season for which outputs are desired.}
}
\value{
A matrix with each entry representing the expected win probability
of the home team in a matchup of the team in that row at home against the
team in that column on the road.
}
\description{
Predict Win Probabilities between every pair of NBA teams with an Elo model
}
\details{
Note that diagonal entries of the return matrix are
not exactly 50% because home-court advantage is taken into account.
}
\examples{
\dontrun{
data(eloLogs2017)
buildWinProbMatrixElo(eloLogs2017, 2017)
}

}
\seealso{
\code{\link{buildPtDiffMatrixElo}} for obtaining a similar matrix
with predicted point differentials, and \code{\link{buildModelsElo}} to obtain
both types of matrices at once.
}
\author{
astroud
}
