% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGamelogColumnsBT.R
\name{checkGamelogColumnsBT}
\alias{checkGamelogColumnsBT}
\title{Check whether the columns of a Game Log are valid}
\usage{
checkGamelogColumnsBT(gameLog)
}
\arguments{
\item{gameLog}{The game log to be supplied to the \code{gameLog} argument
of either \code{\link{buildWinProbMatrixBT}}, \code{\link{buildPtDiffMatrixBT}}, or
\code{\link{buildModelsBT}}.}
}
\value{
Either returns nothing, or errors out.
}
\description{
Checks whether the column names match up with the column names expected
by the functions within this package that fit Bradley-Terry models.
}
\author{
astroud
}
\keyword{internal}
