% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{createEloModel}
\alias{createEloModel}
\title{Create Elo models for a certain season using FiveThirtyEight gamelogs}
\usage{
createEloModel(eloLogs, season)
}
\arguments{
\item{eloLogs}{A dataframe with certain columns, for a list see the parameter
\code{eloLogs} of the function \code{\link{buildModelsElo}}. The output of a call
to \code{\link{getEloGameLogs}} will work here.}

\item{season}{An integer indicating the NBA season on which to create these models.}
}
\value{
A list with coefficients for the win-percentage logistic regression, the point
differential linear regression, and each team's final Elo rating for the season (see Details).
}
\description{
Create Elo models for a certain season using FiveThirtyEight gamelogs
}
\details{
Uses logistic regression to predict a home team's chances of winning
given the difference between their Elo rating and their opponent's. Uses linear
regression to predict a home team's expected point advantage given the difference
between their Elo rating and their opponent's.
}
\author{
astroud
}
\keyword{internal}
