% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloLogs2017.R
\docType{data}
\name{eloLogs2017}
\alias{eloLogs2017}
\title{Game logs with Elo ratings from the 2017 NBA season}
\description{
A dataset with one row for every game of the 2017 NBA season.
Includes pre- and post-game elo ratings for both teams,
FiveThirtyEight's new CARMELO ratings, and each team's pregame
win probabilities according to both systems.
Note that this dataset includes both regular season and
postseason data; use the attribute \code{playoff} to filter this.
}
\keyword{data}
