% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameLogs.R
\name{extractGameLog}
\alias{extractGameLog}
\title{Extract game logs from a stats.nba.com API response}
\usage{
extractGameLog(htmlGamelog)
}
\arguments{
\item{htmlGamelog}{An HTML nodeset corresponding to a response from the
stats.nba.com teamgamelog API endpoint.}
}
\value{
A data frame containing all the gamelogs from a single
stats.nba.com API response.
}
\description{
Extract game logs from a stats.nba.com API response
}
\author{
astroud
}
\keyword{internal}
