% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelElo538.R
\name{extractTeamElos}
\alias{extractTeamElos}
\title{Extract team Elo ratings for a given NBA season}
\usage{
extractTeamElos(eloLogs, season)
}
\arguments{
\item{eloLogs}{A dataframe with certain columns, for a list see the parameter
\code{eloLogs} of the function \code{\link{buildModelsElo}}. The output of a call
to \code{\link{getEloGameLogs}} will work here.}

\item{season}{An integer indicating the NBA season from which to extract Elo ratings.}
}
\value{
A data frame with each team and their final Elo rating.
}
\description{
Extract team Elo ratings for a given NBA season
}
\examples{
\dontrun{
eloLogs = getEloGameLogs(2015, 2017)
extractTeamElos(eloLogs, 2017)
}

}
\author{
astroud
}
