% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelogs2017_full.R
\docType{data}
\name{gamelogs2017_full}
\alias{gamelogs2017_full}
\title{Game logs with full statistics from the 2017 NBA season}
\description{
A dataset with one row for every game of the 2017 NBA season.
Includes win-loss record, field goals, three-pointers, and
free-throws made and attempted, offensive, defensive and total
rebounds, assists, steals, blocks, turnovers, fouls, and points
for both the home and away teams, along with identifiers for
each game itself.
}
\keyword{data}
