% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelogs2017_simple.R
\docType{data}
\name{gamelogs2017_simple}
\alias{gamelogs2017_simple}
\title{Simplified game logs from the 2017 NBA season}
\description{
A dataset with one row for every game of the 2017 NBA season.
Includes only game ID, game date, the two teams playing, and
their point totals, for ease in experimenting with models
that ignore other attributes.
}
\keyword{data}
