% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameLogs.R
\name{getGameLogs}
\alias{getGameLogs}
\title{Obtain NBA game logs}
\usage{
getGameLogs(from = 2017, to = from, type = c("Regular Season",
  "Playoffs"), simple = FALSE)
}
\arguments{
\item{from}{An integer denoting the first season from which to search.
As NBA seasons span two calendar years, use the later of the two years
as the passed-in parameter (e.g. for the 2017-18 season, use 2018).}

\item{to}{An integer denoting the last season to include in the search.}

\item{type}{A character string specifying whether regular season game logs
or playoff game logs are desired.}

\item{simple}{A logical value indicating whether the user would like simple
game logs (only gameID, date, teams, and score) or full game logs (all recorded
statistics for both teams).}
}
\value{
A data frame with game logs of the given \code{type} for every NBA
season between \code{from} and \code{to}, inclusive.
}
\description{
Gather gamelogs with either scores-only or comprehensive data for a
specified range of NBA seasons.
}
\details{
Note that a small delay is included between each API request. Because
stats.nba.com only makes gamelogs available for a specific team/season/type
combination, each season of gamelogs gathered will require 30 API requests, and
thus scraping gamelogs for the entire history of the NBA will take some time. As
far as I am aware, the stats.nba.com API has no volume limits, but in tests
without a fraction-of-a-second delay, the site refused to respond after an
excessive rate/quantity of requests, so the delay was included.
}
\examples{
\dontrun{
getGameLogs()
getGameLogs(2018, type = "Regular Season")
getGameLogs(2000, 2017, type = "Playoffs")
getGameLogs(1947, 2017, simple = TRUE)
}

}
\seealso{
\code{\link{getEloGameLogs}} for collecting game logs with Elo ratings
for each team, and \code{\link{getSeasonStats}} for gathering team statistics
aggregated over the course of a season. \code{\link{buildModelsBT}} can be useful
for modeling team strength using these game logs.
}
\author{
astroud
}
