% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameLogs.R
\name{getQueryParamMatrix}
\alias{getQueryParamMatrix}
\title{Create a matrix of parameters for nba.stats.com queries}
\usage{
getQueryParamMatrix(from, to, type)
}
\arguments{
\item{from}{A year, integer or string -- the first year
from which data is desired.}

\item{to}{A year, integer or string -- the last year from
which data is desired.}

\item{type}{Either "Playoffs" or "Regular Season".}
}
\value{
A data frame where each row contains query parameters
to send to \code{stats.nba.com/teamgamelog}, and where each
column contains the parameter names.
}
\description{
Create a matrix of parameters for nba.stats.com queries
}
\author{
astroud
}
\keyword{internal}
