% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonStats.R
\name{getSeasonStats}
\alias{getSeasonStats}
\title{Obtain NBA season statistics}
\usage{
getSeasonStats(from = 2017, to = from, type = c("per game", "totals",
  "per 100 poss"))
}
\arguments{
\item{from}{An integer denoting the first season from which to search.
As NBA seasons span two calendar years, use the later of the two years
as the passed-in parameter (e.g. for the 2017-18 season, use 2018).}

\item{to}{An integer denoting the last season to include in the search.}

\item{type}{A character string specifying whether per-game statistics, total
statistics, or per-100-possession statistics (when available) are desired.}
}
\value{
A data frame with team statistics according to \code{type} for
every NBA season between \code{from} and \code{to}, inclusive.
}
\description{
Returns a data-frame containing various season summary statistics for
every NBA team in each specified season. Statistics include metrics
for a team itself and for that team's opponents. Throws an error if a
season does not exist in the basketball-reference database.
}
\examples{
getSeasonStats()
getSeasonStats(2018, type = "per 100 poss")
\dontrun{
getSeasonStats(2000, 2017, type = "totals")
}

}
\seealso{
\code{\link{getGameLogs}} for gathering individual game data, and
\code{\link{getEloGameLogs}} for gathering game logs with elo ratings.
}
\author{
astroud
}
