% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonStats.R
\name{getStatsFromHTML}
\alias{getStatsFromHTML}
\title{Extract team statistics from a Basketball-Reference HTML Table}
\usage{
getStatsFromHTML(statsHTML, year, opponent = FALSE)
}
\arguments{
\item{statsHTML}{An HTML nodeset corresponding to a data table from Basketball-Reference.}

\item{year}{The season to which the data corresponds.}

\item{opponent}{A logical value indicating whether these statistics record a team's own
performance or the performance of teams playing against them.}
}
\value{
A data frame containing the statistics in \code{statsHTML}.
}
\description{
Extract team statistics from a Basketball-Reference HTML Table
}
\author{
astroud
}
\keyword{internal}
