% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonStats.R
\name{getTeamStats}
\alias{getTeamStats}
\title{Extract team statistics from a Basketball-Reference season summary webpage}
\usage{
getTeamStats(seasonSummaryHTML, type = "per 100 poss")
}
\arguments{
\item{seasonSummaryHTML}{An HTML nodeset corresponding to a Basketball-Reference.com
season summary webpage.}

\item{type}{A character string specifying whether per-game statistics, total
statistics, or per-100-possession statistics (when available) are desired.}
}
\value{
A data frame with team statistics and team opponent statistics of the specified
\code{type} for the season contained in \code{seasonSummaryHTML}.
}
\description{
Extract team statistics from a Basketball-Reference season summary webpage
}
\author{
astroud
}
\keyword{internal}
