% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonStats.R
\name{getTeamWL}
\alias{getTeamWL}
\title{Extract win-loss records from a Basketball-Reference season-summary webpage}
\usage{
getTeamWL(seasonSummaryHTML)
}
\arguments{
\item{seasonSummaryHTML}{An HTML nodeset corresponding to a Basketball-Reference.com
season summary webpage.}
}
\value{
A data frame containing win-loss records for every NBA team during the
season of information contained in \code{seasonSummaryHTML}.
}
\description{
Extract win-loss records from a Basketball-Reference season-summary webpage
}
\author{
astroud
}
\keyword{internal}
