% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonStats.R
\name{getYear}
\alias{getYear}
\title{Extract the year from a Basketball-Reference season-summary webpage}
\usage{
getYear(html)
}
\arguments{
\item{html}{An HTML nodeset corresponding to a Basketball-Reference.com
season summary webpage.}
}
\value{
A character string containing the year corresponding to the
data contained in \code{html}.
}
\description{
Extract the year from a Basketball-Reference season-summary webpage
}
\author{
astroud
}
\keyword{internal}
