% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffSeries.R
\name{playoffSeries}
\alias{playoffSeries}
\title{Bar plot predicting playoff series}
\usage{
playoffSeries(prob, home, away, plot = FALSE)
}
\arguments{
\item{prob}{A probability matrix with each entry representing the
probability of the team in that row defeating the tean in that
column. It can be generated by \code{\link{buildWinProbMatrixBT}} or
\code{\link{buildWinProbMatrixElo}}.}

\item{home}{The team with home-court advantage in the series.}

\item{away}{The team without home-court advantage.}

\item{plot}{A logical value indicating whether the user would
like to generate a barplot of series outcomes.}
}
\value{
A numeric vector giving the probability of each playoff series outcome.
}
\description{
Bar plot predicting playoff series
}
\details{
Note that while both the Elo model and Bradley-Terry model
use three-letter abbreviations to refer to teams, these abbreviations
differ because the models are built using datasets from different sources
(FiveThirtyEight versus stats.nba.com). Thus, be careful when specifying
team names. We hope to add a function in the future that will allow for
easier conversion between team names and abbreviations.
}
\examples{
playoffSeries(seasonWinProb2017, home = "BOS", away = "SAS")
}
\author{
kitliu5 and astroud
}
