% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffSeries.R
\name{plotSeriesOutcomes}
\alias{plotSeriesOutcomes}
\title{Bar plot predicting playoff series outcomes}
\usage{
plotSeriesOutcomes(home, away, outcomes)
}
\arguments{
\item{home}{The team with home-court advantage in the series.}

\item{away}{The team without home-court advantage.}

\item{outcomes}{A vector with all possible outcomes of a 7-game
playoff series (e.g. Home wins in 6), each with its associated probability.}
}
\value{
A bar plot showing the probability of each playoff series outcome.
}
\description{
Bar plot predicting playoff series outcomes
}
\details{
This barplot is currently a fairly poor visualization - one of our
primary TODOs is to clean up and prettify the visualization it gives. We plan
on learning ggplot to implement this.
}
\author{
kitliu5
}
\keyword{internal}
