% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSeasonWins.R
\name{projectSeasonWins}
\alias{projectSeasonWins}
\title{Estimate season win totals for each team in a win-probability matrix}
\usage{
projectSeasonWins(winProbs)
}
\arguments{
\item{winProbs}{A probability matrix with each entry representing the
probability of the team in that row defeating the team in that
column. It can be generated by \code{buildWinProbMatrixBT} or
\code{buildWinProbMatrixElo}.}
}
\value{
A named numeric vector where each entry's name is a team and
the entry is their projected number of wins.
}
\description{
Estimate season win totals for each team in a win-probability matrix
}
\details{
Note that as currently implemented, this function does not
account for the differing number of games played against each team,
but does take into account an even split between home and road games.
This function also assumes an 82-game season as is the standard in the
modern NBA, so if the probability matrix is generated on a season that
had fewer games, this function will extrapolate out to a full 82.
}
\examples{
\dontrun{
projectSeasonWins(buildWinProbMatrixBT(gamelogs2017_full))
projectSeasonWins(buildWinProbMatrixElo(eloLogs2017))
}

}
\author{
astroud
}
