% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameLogs.R
\name{queryNBAStatsAPI}
\alias{queryNBAStatsAPI}
\title{Query the stats.nba.com API to obtain game logs}
\usage{
queryNBAStatsAPI(TeamID, Season, SeasonType)
}
\arguments{
\item{TeamID}{An integer corresponding to the stats.nba.com team ID.}

\item{Season}{A character string of the form YYYY-YY.}

\item{SeasonType}{Either "Regular Season" or "Playoffs".}
}
\value{
An html nodeset containing a stats.nba.com API response.
}
\description{
Query the stats.nba.com API to obtain game logs
}
\details{
This function uses the "teamgamelog" API endpoint.
An example URL that will return a response with regular season
game logs for the Atlanta Hawks with all games from 2014-15 is
\url{http://stats.nba.com/stats/teamgamelog/?TeamID=1610612737&Season=2014-15&SeasonType=Regular\%20Season}.
}
\author{
astroud
}
\keyword{internal}
