% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonPtDiff2017.R
\docType{data}
\name{seasonPtDiff2017}
\alias{seasonPtDiff2017}
\title{A point differential matrix with all possible matchups between NBA teams in
2017. For each entry, the team of that row is at home court while the team
of that column is on road. Note that diagonal entries are just estimates of
the home court advantage coefficient. The column names and the first column
of the matrix are names of all NBA teams in regular season of 2017.}
\description{
A point differential matrix with all possible matchups between NBA teams in
2017. For each entry, the team of that row is at home court while the team
of that column is on road. Note that diagonal entries are just estimates of
the home court advantage coefficient. The column names and the first column
of the matrix are names of all NBA teams in regular season of 2017.
}
\keyword{data}
