% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonStats2017_per100p.R
\docType{data}
\name{seasonStats2017_per100p}
\alias{seasonStats2017_per100p}
\title{A data frame with per-100-possession NBA team season summary data from 2017}
\description{
This data frame contains one row for each team in the 2016-17
NBA season. Features of this dataset include win-loss record,
shooting statistics, rebounds, assists, blocks, steals, fouls,
and points, averaged across every 100 possessions for each
team.
}
\keyword{data}
