% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonStats2017_perGame.R
\docType{data}
\name{seasonStats2017_perGame}
\alias{seasonStats2017_perGame}
\title{A data frame with per-game team season summary data from the 2017 NBA season}
\description{
This data frame contains one row for each team in the 2016-17
NBA season. Features of this dataset include win-loss record,
shooting statistics, rebounds, assists, blocks, steals, fouls,
and points, averaged out on a per-game basis.
}
\keyword{data}
