% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonWinProb2017.R
\docType{data}
\name{seasonWinProb2017}
\alias{seasonWinProb2017}
\title{A probability matrix in which each entry represents the
probability of the NBA team in that row defeating the NBA
team in that column in regular season of 2017}
\description{
The rownames and the first column of the matrix are names
of all NBA teams in regular season of 2017. Diagonal entries
indicate the effect of home-court advantage when two identical
teams play against each other.
}
\keyword{data}
