% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playoffSeries.R
\name{seriesOutcomes}
\alias{seriesOutcomes}
\title{Predicting playoff series results}
\usage{
seriesOutcomes(teamH, teamA, t1pH, t1pA)
}
\arguments{
\item{teamH}{The team with home-court advantage (HCA) for the series.}

\item{teamA}{The team without home-court advantage.}

\item{t1pH}{Probability of the team with HCA winning at home.}

\item{t1pA}{Probability of the team with HCA winning, when they play on the road.}
}
\value{
A numeric vector with probabilities of each possible series outcome.
}
\description{
Predicting playoff series results
}
\author{
astroud
}
\keyword{internal}
